port json5 serializer, move to qup, docs
