+++ src/main/java/org/quiltmc/config/api/serializer/TomlSerializer.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2023 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.quiltmc.config.api.serializer;
 
 import com.electronwill.nightconfig.core.CommentedConfig;
@@ -27,10 +43,17 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * A default serializer that writes in the <a href="https://toml.io/en/">TOML format</a>.
+ */
 public final class TomlSerializer implements Serializer {
 	public static final TomlSerializer INSTANCE = new TomlSerializer();
 	private final ConfigParser<CommentedConfig> parser = new TomlParser();
 	private final ConfigWriter writer = new TomlWriter();
+
+	private TomlSerializer() {
+
+	}
 
 	@Override
 	public String getFileExtension() {
